src/requirements.txt
absl-py==0.7.1
astor==0.7.1
cachetools==3.1.1
certifi==2019.3.9
chardet==3.0.4
cloudpickle==0.8.0
cycler==0.10.0
cytoolz==0.9.0.1
dask==1.1.4
decorator==4.4.0
gast==0.2.2
google-auth==1.7.0
google-auth-oauthlib==0.4.1
google-pasta==0.1.7
GPUtil==1.4.0
grpcio==1.19.0
h5py==2.9.0
hdf5storage==0.1.15
idna==2.8
imageio==2.5.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
llvmlite==0.28.0
Markdown==3.1
matplotlib==3.0.3
mkl-fft==1.0.10
mkl-random==1.0.2
mock==2.0.0
networkx==2.2
numba==0.43.1
numpy==1.16.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.0.0.21
opt-einsum==3.1.0
pbr==5.1.3
Pillow==6.0.0
protobuf==3.7.1
pyasn1==0.4.7
pyasn1-modules==0.2.7
pyparsing==2.3.1
pyreadline==2.1
python-dateutil==2.8.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
tensorboard==1.12.2
tensorflow-estimator==2.0.1
tensorflow-gpu==1.12.0
termcolor==1.1.0
toolz==0.9.0
tornado==6.0.2
tqdm==4.31.1
urllib3==1.25.6
Werkzeug==0.15.2
wincertstore==0.2
wrapt==1.11.2
